solver works sometimes--
Really, there are situations where we can get stuck repeating unsolvable
choices, that I have yet to note correctly.

So the choice is to just abort with an error state.

Theres a lot of guessing, and I have been too lazy to actually work through
it. Thats the approach of brute force and womp womp.

TODO: make this not suck